refactor: Sponsor Info 컴포넌트 화 및 디자인 반영
