+++ components/sponsor/index.ts
@@ -1,0 +1,1 @@
+export { default as SponsorInfoInputBox } from './SponsorInfoInputBox';
